+++ src/main/java/toy/board/service/cache/CacheService.java
@@ -1,0 +1,8 @@
+package toy.board.service.cache;
+
+public interface CacheService {
+    void setValues(String s, String authCode, long authCodeExpirationMillis);
+
+    boolean deleteIfValueExistAndEqualTo(String s, String authCode);
+
+}
